refactor: extract reducer focus state
